refactor link component
