+++ src/pages/index.tsx
@@ -1,9 +1,6 @@
 import React from "react"
-import { useTranslation } from "gatsby-plugin-react-i18next"
 import { Container, Row, Column } from "@components/layout"
 import Header from "@components/patterns/header"
-import HamburgerButton from "@components/button/HamburgerButton"
-import { Text } from "@chakra-ui/react"
 
 const Home = (): React.ReactElement => {
   return (
